backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
pyrogram==2.0.106
TgCrypto==1.2.5
websockets==12.0
httpx==0.26.0
textblob==0.18.0
aiofiles==23.2.1
Pillow==10.2.0
